refactor: 将主键类型从 Long 更改为 Integer，优化数据模型和数据库访问层

+++ avalon-application/avalon-admin-server/src/main/java/io/github/lishangbu/avalon/admin/service/dataset/impl/BerryFlavorServiceImpl.java
@@ -37,7 +37,7 @@ public class BerryFlavorServiceImpl implements BerryFlavorService {
         berryFlavorData -> {
           BerryFlavor berryFlavor = new BerryFlavor();
           berryFlavor.setInternalName(berryFlavorData.name());
-          berryFlavor.setId(berryFlavorData.id().longValue());
+          berryFlavor.setId(berryFlavorData.id());
           berryFlavor.setName(berryFlavorData.name());
           LocalizationUtils.getLocalizationName(berryFlavorData.names())
               .ifPresent(name -> berryFlavor.setName(name.name()));
@@ -69,7 +69,7 @@ public class BerryFlavorServiceImpl implements BerryFlavorService {
   }
 
   @Override
-  public void removeById(Long id) {
+  public void removeById(Integer id) {
     berryFlavorRepository.deleteById(id);
   }
 }
