+++ avalon-modules/avalon-dataset/avalon-dataset-repository/src/main/java/io/github/lishangbu/avalon/dataset/repository/ItemCategoryRepository.java
@@ -13,5 +13,5 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface ItemCategoryRepository
-    extends ListCrudRepository<ItemCategory, Long>,
-        ListPagingAndSortingRepository<ItemCategory, Long> {}
+    extends ListCrudRepository<ItemCategory, Integer>,
+        ListPagingAndSortingRepository<ItemCategory, Integer> {}
